cleaned up login and register by splitting them in seperate jsx files and added comments to everything
