+++ friendsshelvesreactconversion/src/components/BookShow.jsx
@@ -5,6 +5,7 @@ import {useState} from 'react';
 
 function BookShow({book, onDelete, onEdit}){
 
+    //Handle the edit menu for every book
     const[showEdit, setShowEdit] = useState(false);
 
     const handleDeleteClick = () =>{
@@ -15,21 +16,25 @@ function BookShow({book, onDelete, onEdit}){
             setShowEdit(!showEdit);
     };
 
+    //Handle submissions for edits
     const handleSubmit = (id, newTitle, newAuthor, newISBN, newImage, newBlurb) => {
 
         onEdit(id, newTitle, newAuthor, newISBN, newImage, newBlurb);
         setShowEdit(false);
     }
 
+    //Show the book as object
     let content = <div><p>{book.title}</p>{book.author}<p>{book.ISBN}</p>{book.blurb}</div>
     ; 
 
+    //if edit button has been pressed, show edit menu for set book
     if (showEdit){
         content = <BookEdit onSubmit={handleSubmit} book={book}/>;
     }
 
-    let bookImage = 'https://covers.openlibrary.org/b/isbn/'.concat(book.ISBN,'-M.jpg');
+    // API for book images, might be used later: let bookImage = 'https://covers.openlibrary.org/b/isbn/'.concat(book.ISBN,'-M.jpg');
     
+    //Show book or edit menu for each book
     return <div className = "book-show">
             {content}
                 <div ClassName = "actions">
