+++ friendsshelvesreactconversion/src/components/LoginRegister/LoginForm.jsx
@@ -1,0 +1,67 @@
+import {useState} from 'react'; 
+
+
+function LoginForm ({onSubmit}) {
+
+     //Handle the data from the login form
+
+     const [formData, setFormData] = useState({username: "", password: ""})
+
+     const handleChange = (event) => {
+         const changedField = event.target.name;
+         const newValue = event.target.value;
+         setFormData(currData => {
+             currData[changedField] = newValue;
+             return{...currData};
+         })
+ 
+     }
+ 
+     const handleSubmit = (event) => {
+         event.preventDefault();
+         onSubmit(formData.username, formData.password);
+     };
+ 
+     //Login form
+
+    return (<div>
+        <h2>Login</h2>
+            <form onSubmit = {handleSubmit}>
+                <p>
+                    <label>Username </label>
+                    <input 
+                    type = "text"
+                    placeholder = "Username" 
+                    className="input" 
+                    value = {formData.username} 
+                    onChange = {handleChange}
+                    name = "username"
+                    id="username"
+                    />
+                </p>
+                <p>
+                    <label>Password </label>
+                    <input
+                    type = "password"
+                    placeholder = "Password" 
+                    className="input" 
+                    value = {formData.password} 
+                    onChange = {handleChange}
+                    name = "password"
+                    id="password"
+                    />
+                </p>
+                <button onClick={handleSubmit} className="button">Login</button>
+            </form>
+    </div>)
+
+}
+
+export default LoginForm;  
+
+
+
+
+
+
+
